# Improve performance of memoized validator functions.