The memoize cache key was based on a hash of the entire graph. This hash was being generated each time a link was validated. On my system, validating each link took around 100ms with a graph of around 2000 links. With this change it appears that validating each link takes less than 1ms.

Of course, we don't want cache key collisions. Do you think using `link.hash` and `graph.root` should be unique enough for the expected use cases?