@HerbCaudill 2024-09-12
Hi, Lucas - thanks for all these PRs! Sorry for not responding earlier - I was afk for the past month or so. I'll go through these now.

I'd love to see your benchmarks and maybe add them to the repo - would that code be easy for you to pull up?