@leblowl 2024-09-12
You're welcome. I've been testing end-to-end by adding console.log statements in the code and then running a scenario. Would that test code setup be something you are interested in? I could probably come up with some unit tests as well.