@HerbCaudill 2024-09-12
vitest has [built-in benchmarking support](https://vitest.dev/guide/features#benchmarking) using tinybench - could be cool to simulate some of these scenarios within the test suite as a way of documenting these performance improvements, and also providing a baseline for further optimization work.