@leblowl 2024-09-12
I just added a benchmark. How does that look to you?

Before PR:
```
 ✓ packages/auth/src/test/auth.benchmark.ts (1) 47002ms
   ✓ auth (1) 47000ms
     name                      hz       min       max      mean       p75       p99      p995      p999     rme  samples
   · a new member joining  0.3610  2,711.03  2,829.56  2,770.39  2,814.50  2,829.56  2,829.56  2,829.56  ±1.05%       10
```

After PR:
```
 ✓ packages/auth/src/test/auth.benchmark.ts (1) 4631ms
   ✓ auth (1) 4629ms
     name                      hz     min     max    mean     p75     p99    p995    p999      rme  samples
   · a new member joining  3.6959  236.43  367.57  270.57  282.42  367.57  367.57  367.57  ±10.20%       10
```