@leblowl 2024-09-13
It looks like some tests are failing due to the mutable nature of the graph. For example, in this test: https://github.com/local-first-web/auth/blob/6ab7e9c2947a9d2e865fe5fa8e9967f86f887d5c/packages/crdx/src/validator/test/validate.test.ts#L49-L50

We modify the root entry, but the entry's hash doesn't change. So given this, the memoization changes in this PR won't work.

One option is to remove memoization on validation. It looks like validation is only used when creating the Team and then when receiving a message: https://github.com/local-first-web/auth/blob/6ab7e9c2947a9d2e865fe5fa8e9967f86f887d5c/packages/crdx/src/sync/receiveMessage.ts#L73

And upon receiving a message, if the graph changes at all, the original memoization approach isn't going to help because ```hash('memoize', graph)``` will change.

Another option is to keep the original memoization, but compute the hash of the graph only once before we call the individual validators: https://github.com/local-first-web/auth/blob/6ab7e9c2947a9d2e865fe5fa8e9967f86f887d5c/packages/crdx/src/validator/validate.ts#L60

So that we are not computing the hash of the graph when validating each link. In this scenario we would still need to compute the hash of each link. But if the graph ever changes because of a new message, from what I can tell, the memoization will not help and will increase memory usage.