@HerbCaudill 2024-09-13
Yeah - I've done some experiments and I think we should memoize the `validate` function rather than each of the individual validators. This gives us performance comparable to what you were getting and gets the tests passing again. I'll push some changes for you to review.