@leblowl 2024-09-13
Sounds good. Unfortunately, since this PR is using the `TryQuiet/auth` repo, it doesn't look like I have the option to allow changes by maintainers. I opened another PR from my personal repo that should allow edits: https://github.com/local-first-web/auth/pull/129